requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pydantic
alembic
python-dotenv
passlib[bcrypt]
PyJWT
python-multipart
bcrypt